chore: refactor blockchain providers
